Add JSON data filtering documentation
This adds documentation for an upcoming feature for in the Terraform
Cloud UI that allows for the display of lower-level policy check results
in JSON format, with the ability to filter these results using a
jq-subset filter. This feature is also available to the state viewer,
where it can be used to filter data in the Terraform state.

Once the feature is launched, the intention is to update this
documentation with screenshots.