Delete ALL remaining provider submodules and symlinks (#1573)
The one exception is the symlink to the `terraform` provider and its
`terraform_remote_state` data source in the Terraform core repo. That one is
handled specially, and gets presented as part of the language docs.

This change must not be merged to master until all of the remaining providers
have been published and redirected to the registry.