Update description and note per PR comments
Change the into to more clearly show that while SDKv2 is the currently
preferred SDK for now, it will eventually be replaced by the framework.
Remove recommendation to stick with SDKv2.
But still warn users that framework is in early development and point them
toward the Which SDK? page that can help them make the right choice for their
provider.