Update About the Docs page content
Goal is to make the content more concise (it was previously very text-heavy)
and also to help new users understand where they should go next. When you don't
know the products or what they're used for, it can be very difficult to know where
to start or how to find the right docs for your usecase. The new content tries to
point users in the right direction and match content with what they are trying to do
(or what they may want to do) with Terraform.