Refine Plugin Framework Docs  (#1782)
* Update plugin development homepage with new framework learn tutorial

* Update learn tutorial links in sidebar

Goal is to make it clearer what these links are and why a user would want to click them.

* Change Accessing title to match 'ing' form in sidebar

* Update language on Publishing to Registry page

Make introduction more concise and active. Add numbered steps below to make it clear that there
are two separate things developers need to do when adding a manifest.

* Use tables on attributes page

Makes the content more readable and reduces page length.

* Update Create Types and Values title to remove extra words

* Add homepage for SDKv2

The home should no longer point to the plugin home. This may confuse users.
The new overview page parallels the one for the plugin framework and also
helps users understand which topics they may want to visit next.

* Update plugin development homepage to link to new sdkv2 overview page

* Update plugin framework overview page with more detail

Goal is to make this page a bit more useful to users by helping them understand
where they should go next.

* Fix broken schemas link on sdkv2 overview page

* Update SDKv2 link on framework overview to point to sdkv2 home

* Update sdkv2 link on which sdk page to point to sdkv2 home

* Update content/source/docs/extend/sdkv2-intro.html.md

Co-authored-by: Paddy <paddy@hashicorp.com>

* Update content/source/docs/plugin/framework/types.html.md

Co-authored-by: Paddy <paddy@hashicorp.com>

* Update description and note per PR comments

Change the into to more clearly show that while SDKv2 is the currently
preferred SDK for now, it will eventually be replaced by the framework.
Remove recommendation to stick with SDKv2.
But still warn users that framework is in early development and point them
toward the Which SDK? page that can help them make the right choice for their
provider.

* Update Which SDK intro for concision and to add bullets

* Update content/source/docs/extend/sdkv2-intro.html.md

Co-authored-by: Paddy <paddy@hashicorp.com>

* Update content/source/docs/extend/index.html.md

Co-authored-by: Aareet Mahadevan <aareet@users.noreply.github.com>

* Update content/source/docs/plugin/framework/publishing.html.md

Co-authored-by: Aareet Mahadevan <aareet@users.noreply.github.com>

* Address PR feedback on which SDK

* Address code comments about version example

* Update version to match name example

* Add plugin overview image

* Add overview image to plugin development home

* Center Terraform icon

* Update version number (again) and include explanation about manifest version

Co-authored-by: Paddy <paddy@hashicorp.com>
Co-authored-by: Aareet Mahadevan <aareet@users.noreply.github.com>