Redirect docs/plugin/framework/publishing.html to docs/registry/providers/publishing.html
Reference: https://github.com/hashicorp/terraform-plugin-framework/issues/87

The manifest file is now recommended for any provider release so it is now included in the general provider publishing information. Each provider should include the appropriate protocol version (5.0 for Terraform Plugin SDK version 2 and 6.0 for Terraform Plugin Framework).

The Publishing to the Registry sidebar is promoted from under Framework to the top level Plugin Development navigation, as it is appliable to either SDK.