language: Sub-pages of "upgrade-guides" now only on the v1.1 branch
Now that we have the possibility of viewing docs from older version
branches on the live site, it's confusing that e.g. being on the v1.3
upgrade guide and then choosing "v1.2" from the version selector ends
up at an invalid URL (the non-existing v1.3 guide on the v1.2 branch)
instead of at the v1.2 upgrade guide.

We're going to fix this by making the v1.2 and subsequent release branches
include only the upgrade guide for their own minor release, and always
at the same URL so that the version switcher chooses between the different
version upgrade guides.

As a special case, the v1.1 branch continues to mean "v1.1 and earlier"
for documentation purposes and so the older upgrade guides will live on
there. In support of that, this introduces some redirects from where those
pages used to be to the corresponding locations under the v1.1 branch,
which will be their canonical home moving forward.

To avoid complexities of names changing when we switch between versions,
the v1.2 and later release branches will have their single upgrade guide
published at the plural URL "upgrade-guides" so that it will match with
the Overview page on the v1.1 branch. This is a _little_ quirky but a
small price to pay to make the version selector work consistently across
all versions.